LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.scale.nonfinite.worker-expected.txt


PASS scale() with Infinity/NaN is ignored
